% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bout.R
\name{berm}
\alias{berm}
\title{Biexponential refractory model, BERM}
\usage{
berm(irt, delta)
}
\arguments{
\item{irt}{numeric, the inter-response times}

\item{delta}{numeric, the refractory period}
}
\value{
a data frame with the parameters \eqn{a, b, q} where
\eqn{a, b} are the mean within and between bout IRTs and \eqn{q} is the proportion of responses
that are bouts, and \eqn{\delta} is the refractory period.
}
\description{
Implements the biexponential refractory model, or berm
\eqn{p(IRT < \tau) = (1-q)e^{-a(IRT-\delta)} + qe^{-b(IRT-\delta)}}
where \eqn{a, b} are the mean within and between bout IRTs and \eqn{q} is the proportion of responses
that are bouts, and \eqn{\delta} is the refractory period.
}
