% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bout.R
\name{biexponential}
\alias{biexponential}
\title{Biexponential model for the estimation of within and between bouts}
\usage{
biexponential(irt)
}
\arguments{
\item{irt}{numeric, the inter-response times}
}
\value{
the simple biexponential returns numeric vector of the biexponential
model parameters \eqn{L=1/\theta}, the bout length,
\eqn{W = 1/w}, the within bout rate, and \eqn{B = 1/b} the bout initiation rate, where
\eqn{\theta, w, b} are the proportion of responses that are bouts, the mean within bout IRT, and
the mean between-bout IRT.
The biexponential refractory model returns a data frame with the parameters \eqn{a, b, q} where
\eqn{a, b} are the mean within and between bout IRTs and \eqn{q} is the proportion of responses
that are bouts, and \eqn{\delta} is the refractory period.
}
\description{
Biexponential model for the estimation of within and between bouts
}
\details{
Implements the finite mixtures of two exponentials
\eqn{p(IRT < \tau) = \theta w e^{-w IRT} + (1-\theta)b e^{-b IRT}}

and the biexponential refractory model, or berm, which is a special case of the biexponential model
}
\examples{
l1 <- 1/0.5
l2 <- 1/0.1
theta <- 0.5
irt <- c(
 rexp(100, l1),
rexp(100, l2)
)
biexponential(irt)
berm(irt, 0.1)
}
