% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_dot_plot.r
\name{box_dot_plot}
\alias{box_dot_plot}
\title{Custom box plot with dots of y ~ x on top of boxplot and stripchart functions}
\usage{
box_dot_plot(
  variable,
  groups,
  xlab = "",
  ylab = "",
  col_pal,
  sort_col_pal = T,
  fill = "white",
  lty = 1,
  boxlty = 1,
  whisklty = 2,
  box_around = T
)
}
\arguments{
\item{variable}{Numeric}

\item{groups}{chr or factor}

\item{xlab}{chr, the name of x axis}

\item{ylab}{crh, the name of y axis}

\item{col_pal}{chr vector, colors of the boxes; if none, a color palette is provided by Polychrome}

\item{sort_col_pal}{logical, should the col_pal be sorted?}

\item{fill}{chr, a color to fill the box}

\item{lty}{integer or chr, linetype of the median (1 by default)}

\item{boxlty}{integer or chr, linetype of the box  (1 by default)}

\item{whisklty}{integer or chr, linetype of the whiskers  (2 by default, dotted)}

\item{box_around}{logical, whether to draw a box or just the axes}
}
\description{
Custom box plot with dots of y ~ x on top of boxplot and stripchart functions
}
\examples{
rxnx <- data.frame(var = rnorm(30), grupo = rep(c("A", "B", "C"), each = 10))
box_dot_plot(rxnx$var, rxnx$grupo, col_pal = "none", lty = 1, boxlty = 2, box_around = FALSE)
}
