% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_km.R
\name{bp_km}
\alias{bp_km}
\title{Break point using kmeans}
\usage{
bp_km(r_times, min_x, max_x, bin_res)
}
\arguments{
\item{r_times, }{numeric, the times at which a response was emitted in a trial}

\item{min_x, }{numeric, the minimal value of x}

\item{max_x, }{numeric, the maximal value of x}

\item{bin_res, }{numeric, the bin resolution}
}
\value{
data.frame
}
\description{
Break point using kmeans
}
\examples{
data("r_times")
bps_km <- bp_km(r_times, 0, 180, 2)
TODO: Error in dimnames(x) <- dn :
}
