% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_multiple.r
\name{ceiling_multiple}
\alias{ceiling_multiple}
\title{Find the nearest multiple}
\usage{
ceiling_multiple(x, multiple)
}
\arguments{
\item{x}{numeric, the value for which we want to finde a multiple}

\item{multiple}{numeric, the multiple}
}
\value{
the nearest multiple
}
\description{
Find the nearest multiple
}
\examples{
ceiling_multiple(8, 10) # returns 10
ceiling_multiple(12, 10) # returns 20
ceiling_multiple(21, 11) # returns 22
}
