% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_discounting_fit.R
\name{exp_fit}
\alias{exp_fit}
\title{Exponential fit with nls}
\usage{
exp_fit(value, delay, initial_guess, max_iter = 1e+05, scale_offset = 0)
}
\arguments{
\item{value}{A numeric vector of the subjective values (indifference points)}

\item{delay}{A numeric vector of the delays used}

\item{initial_guess}{A numeric value providing an initial start for k}
}
\value{
An object of class nls
}
\description{
Exponential fit with nls
}
\examples{
See the examples of ?hyp_fit
}
