% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_trials_opt.R
\name{ind_trials_obj_fun}
\alias{ind_trials_obj_fun}
\title{Objective function for finding the best fit for individual trials}
\usage{
ind_trials_obj_fun(params, r_times, trial_duration)
}
\arguments{
\item{params}{A vector of parameters to be optimized.}

\item{r_times}{Vector of response times}

\item{trial_duration}{Duration of the trial}

\item{optim_method}{Optimization method. See \code{optim} for details.}
}
\value{
a numeric value representing the sum of areas between the response
rate and the target rate.
}
\description{
This function is used by \code{optim} to find the best fit for
individual trials by minimizing the sum of areas between the response rate
and the target rate. Do not call this function directly.
}
