% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_fi.r
\name{objective_bp}
\alias{objective_bp}
\title{Objective function for the breakpoint optimization algorithm}
\usage{
objective_bp(param, r_times, trial_duration)
}
\arguments{
\item{param}{A numeric value of the breakpoint}

\item{r_times}{A numeric vector of response times}

\item{trial_duration}{A numeric value of the trial duration}
}
\value{
A numeric value representing the sum of areas between the response
rate and the target rate.
}
\description{
Objective function for the breakpoint optimization algorithm
for fixed interval trials. This function is used by \code{optim} to find the
optimal breakpoint. Do not call this function directly.
}
