% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_density.R
\name{sample_from_density}
\alias{sample_from_density}
\title{Sample from a density estimate}
\usage{
sample_from_density(x, n)
}
\arguments{
\item{x}{A numeric variable from a (un)known distribution}

\item{n}{Number of samples to return}
}
\value{
A sample with distribution close to x
}
\description{
Sample from a density estimate
}
\examples{
TODO
}
