% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.r
\name{trapezoid_auc}
\alias{trapezoid_auc}
\title{Area under the curve (AUC)}
\usage{
trapezoid_auc(x, y)
}
\arguments{
\item{x}{A numeric vector of x values}

\item{y}{A numeric vector of y values}
}
\value{
A numeric value of the area under the curve
}
\description{
Calculate the area under the curve (AUC) using the trapezoid
method.
}
\examples{
x_values <- c(0, 1, 2, 3, 4) # Delay times
y_values <- c(1, 0.8, 0.6, 0.4, 0.2) # Discounted values
auc_result <- trapezoid_auc(x_values, y_values)
print(paste("Area Under Curve: ", auc_result))
}
