% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unity_normalization.R
\name{unity_normalization}
\alias{unity_normalization}
\title{Min-max normalization (also feature rescaling)}
\usage{
unity_normalization(x)
}
\arguments{
\item{x}{numeric, vector of values to rescale}
}
\value{
A numeric vector rescaled in the range \eqn{x' \in [0, 1]}
}
\description{
Min-max normalization (also feature rescaling)
}
\examples{
x <- 5:100
x_scaled <- unity_normalization(x)
x_scaled
}
