% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_in_interval.R
\name{val_in_interval}
\alias{val_in_interval}
\title{True value in interval}
\usage{
val_in_interval(df, lowLim, upLim, true.val)
}
\arguments{
\item{true.val}{}
}
\description{
True value in interval
}
