% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YEAB-package.R
\docType{package}
\name{YEAB-package}
\alias{YEAB}
\alias{YEAB-package}
\title{YEAB: Package To Analyze Data From Analysis Of Behavior Experiments}
\description{
This is a colletion of functions aimed to analyze data from behavioral experiments from MED output and others. It also have functions to fit exponential or hyperbolic models from delay discounting tasks, exponential mixtures to IRTs, Gaussian plus ramp model for peak procedures data, etc.
}
\author{
\strong{Maintainer}: Emmanuel Alcala \email{jealcala@gmail.com}

Authors:
\itemize{
  \item Rodrigo Sosa \email{rsosas@up.edu.mx}
}

}
\keyword{internal}
